Split e2e_group setup into separate file and cache base image build (#1321, #1322)
